Organize into packages by placing all classes into duke package
