+++ src/main/java/duke/Deadline.java
@@ -1,15 +1,15 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Deadline extends Task{
     LocalDate date;
     LocalTime time = null;
 
-    //date is entered as "__date__ __time__"
-    public Event(String taskName, String dateTime) throws DukeException {
+    public Deadline(String taskName, String dateTime) throws DukeException {
         super(taskName);
-
         dateTime = dateTime.trim();
         String[] spl = dateTime.split(" ");
         if(spl.length == 1) {
@@ -31,7 +31,7 @@ public class Event extends Task{
     }
 
     public void printDate() {
-        System.out.print("(at: ");
+        System.out.print("(by: ");
         System.out.print(this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")));
         printTime();
         System.out.println(")");
@@ -44,9 +44,8 @@ public class Event extends Task{
     }
 
     @Override public void printTask(){
-        //print task type
-        System.out.print("[E]");
-        //print task done symbol
+        System.out.print("[D]");
+
         if(this.done){
             System.out.print("[X] " + this.taskName + " ");
         } else {
