+++ src/main/java/duke/Todo.java
@@ -1,3 +1,4 @@
+package duke;
 public class Todo extends Task{
 
     public Todo(String taskName){
