+++ src/main/java/duke/Ui.java
@@ -1,3 +1,4 @@
+package duke;
 public class Ui {
     final static String LOGO = " ____        _        \n"
             + "|  _ \\ _   _| | _____ \n"
