+++ text-ui-test/input.txt
@@ -1,4 +1,3 @@
-todo t1
 event e1 /at 1999-01-01 23:30
 deadline d1 /by 1999-09-09 23:12
 event e2 /at 1999-01-01
